// packages/kms-rest-client/src/models/ManagedKeyInfo.ts
export interface ManagedKeyInfo {
  /**
   * Unique identifier for the cryptographic key. Can be null if the key identifier is not provided.
   * @type {string}
   * @memberof ManagedKeyInfo
   */
  kid?: string
  /**
   *
   * @type {Jwk}
   * @memberof ManagedKeyInfo
   */
  key: Jwk
  /**
   *
   * @type {SignatureAlgorithm}
   * @memberof ManagedKeyInfo
   */
  signatureAlgorithm?: SignatureAlgorithm
  /**
   *
   * @type {KeyVisibility}
   * @memberof ManagedKeyInfo
   */
  keyVisibility?: KeyVisibility
  /**
   * X.509 certificate chain associated with the key.
   * @type {Array<string>}
   * @memberof ManagedKeyInfo
   */
  x5c?: Array<string>
  /**
   * A reference or alias to the key in the Key Management Service (KMS).
   * @type {string}
   * @memberof ManagedKeyInfo
   */
  alias: string
  /**
   * The Key Management System (KMS) identifier associated with the key.
   * @type {string}
   * @memberof ManagedKeyInfo
   */
  providerId: string
  /**
   *
   * @type {KeyType}
   * @memberof ManagedKeyInfo
   */
  keyType?: KeyType
  /**
   *
   * @type {KeyEncoding}
   * @memberof ManagedKeyInfo
   */
  keyEncoding?: KeyEncoding
  /**
   * Additional configuration options as key-value pairs.
   * @type {{ [key: string]: string; }}
   * @memberof ManagedKeyInfo
   */
  opts?: { [key: string]: string }
}